requirements.txt
pygame>=2.0.0
Pillow>=9.0.0
pytesseract>=0.3.10
numpy>=1.21.0
requests>=2.28.0
reportlab>=3.6.0

# Writing assistant - spell check & grammar (NEW!)
pyspellchecker>=0.7.0
language-tool-python>=2.7.0

# Optional for advanced summarization
transformers>=4.25.0
torch>=1.13.0


pdf2image>=1.16.0